Issue 439: Incorrect manifest entries are being updated in Manifest.
Use the manifest from the state.

The tests have some problems that are exposed by this change.
Therefore, the tests have been rewritten.
The re-write attempts to stay true to the original design where reasonable.

The tests were passing but were not passing for the correct reasons.
Some of the problems:
- The tests are being configured without a beforeEach, possibly sharing the same data between tests.
- The settings are added throughout the file rather than in a consistent manner.
- The tests are testing against non-existent array indexes on an object (resulting in undefined === undefined tests that pass).
- There is code being executed outside of tests that are used inside of tests (and are not inside a beforeEach).