Issue 249: Directory UI - bottom-nav items (left most) is seen to have shifted left.
The min-width is being applied to the text as well.
This forces the text to have a width that it does not actually span across.
This results in it appearing shifted to the left.

Only apply the min-width to the containing block rather than both the containing block and the text.