Issue 470: Update documentation, complete the reset about, add custom style support.
This finishes the reset about logic I meant to complete but forgot to.

TL-Components has a case where there is a conflict with the TL-Header and the About.
Provide a custom style support as a quick fix to allow TL-Components documentation to weak the About in cases where this is a problem.