Build components with Node 16 or greater.
Add minimum requirement of Node 16 and npm 8.

Upgrade dependencies, where possible.
Some packages are not upgrade due to problems or potential problems.

Switch from `rxjs` to `rxjs-compat`, which is the new location for the old `rxjs` that we are using.
This allows for using the `^` character to get the latest bugfix version for the 6.x releases.

Replace tslint with eslint using `tslint-to-eslint-config` like this:
```
npx tslint-to-eslint-config --prettier --comments
```
This requires removing `angular-tslint-rules`.

Fix bug in `scripts/build-version.js` that prevents building.
The directory path `./dist/bundle` may not exist and needs to be created if it does not exist.

Running this on Node 18 is not easily supported.
An extra step of adding an environment variable might be needed, such as:
```
export NODE_OPTIONS=--openssl-legacy-provider
```

Switching to the new Angular (13+, 14+, etc..) would avoid this but this introduces more problems.